# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Goal Planner 后端接口文档",
        "title": "Goal Planner Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/goals/{id}/generate-plan": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plans"
                ],
                "summary": "生成目标计划",
                "operationId": "goalPlanGenerate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/plan.PlanResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/goals/{id}/plan": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plans"
                ],
                "summary": "获取目标计划",
                "operationId": "goalPlanGet",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/plan.PlanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plans"
                ],
                "summary": "编辑目标计划",
                "operationId": "goalPlanUpdate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "计划参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/plan.UpdatePlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/plan.PlanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plans"
                ],
                "summary": "删除目标计划",
                "operationId": "goalPlanDelete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/goals/{id}/regenerate-plan": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plans"
                ],
                "summary": "重新生成目标计划",
                "operationId": "goalPlanRegenerate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/plan.PlanResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
...
        "/api/plans/{id}/next-step": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plans"
                ],
                "summary": "查询计划执行建议",
                "operationId": "planNextStepGet",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "计划ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/plan.NextStepSuggestionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plans"
                ],
                "summary": "生成计划执行建议",
                "operationId": "planNextStepSuggest",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "计划ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/plan.NextStepSuggestionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "plan.NextStepSuggestion": {
            "type": "object",
            "properties": {
                "checklist": {
                    "description": "Checklist 执行清单",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "next_action": {
                    "description": "NextAction 建议执行动作",
                    "type": "string"
                },
                "reason": {
                    "description": "Reason 建议原因",
                    "type": "string"
                },
                "risk": {
                    "description": "Risk 主要风险",
                    "type": "string"
                },
                "summary": {
                    "description": "Summary 执行建议摘要",
                    "type": "string"
                }
            }
        },
        "plan.NextStepSuggestionResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/plan.NextStepSuggestion"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "plan.Phase": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt 创建时间",
                    "type": "string"
                },
                "description": {
                    "description": "Description 阶段描述",
                    "type": "string"
                },
                "id": {
                    "description": "ID 阶段ID",
                    "type": "integer"
                },
                "plan_id": {
                    "description": "PlanID 所属计划ID",
                    "type": "integer"
                },
                "sort_order": {
                    "description": "SortOrder 阶段顺序",
                    "type": "integer"
                },
                "tasks": {
                    "description": "Tasks 任务列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/plan.Task"
                    }
                },
                "title": {
                    "description": "Title 阶段标题",
                    "type": "string"
                },
                "updated_at": {
                    "description": "UpdatedAt 更新时间",
                    "type": "string"
                }
            }
        },
        "plan.Plan": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt 创建时间",
                    "type": "string"
                },
                "goal_id": {
                    "description": "GoalID 目标ID",
                    "type": "integer"
                },
                "id": {
                    "description": "ID 计划ID",
                    "type": "integer"
                },
                "overview": {
                    "description": "Overview 计划概述",
                    "type": "string"
                },
                "phases": {
                    "description": "Phases 阶段列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/plan.Phase"
                    }
                },
                "title": {
                    "description": "Title 计划标题",
                    "type": "string"
                },
                "updated_at": {
                    "description": "UpdatedAt 更新时间",
                    "type": "string"
                },
                "user_id": {
                    "description": "UserID 用户ID",
                    "type": "integer"
                }
            }
        },
        "plan.PlanResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/plan.Plan"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "plan.UpdatePlanRequest": {
            "type": "object",
            "properties": {
                "overview": {
                    "description": "Overview 计划概述",
                    "type": "string"
                },
                "title": {
                    "description": "Title 计划标题",
                    "type": "string"
                }
            }
        },
...
        "response.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}